--- packages/writing/src/lib/next/index.js
@@ -1,5 +1,0 @@
-const { join } = require("path");
-
-module.exports = {
-  postsDirectory: join(process.cwd(), "src/pages/writing"),
-};
