+++ packages/writing/src/lib/next/index.mjs
@@ -1,0 +1,3 @@
+import { join } from "node:path";
+
+export const postsDirectory = join(process.cwd(), "src/pages/writing");
